requirements.txt
flask
requests
fastapi
tensorflow
pillow
uvicorn
numpy
gunicorn
